requirements.txt
# fastapi==0.111.1
# uvicorn[standard]==0.23.2
# SQLAlchemy==2.0.46
# asyncpg==0.27.0
# pydantic==2.5.2
# python-dotenv==1.0.0
# Pillow==10.4.0
# requests==2.32.3
# pytesseract==0.3.10
fastapi==0.111.1
uvicorn[standard]==0.24.0
SQLAlchemy==2.0.46
asyncpg==0.29.0
pydantic==2.5.2
python-dotenv==1.0.0
Pillow==10.4.0
requests==2.32.3
pytesseract==0.3.10
stripe==8.8.0
pypdf==5.1.0
pdf2image==1.17.0
arq==0.26.3
redis==5.0.8
